// frontend/src/type/admin/adminproduct.type.ts
export interface Product {
    id: number
    name: number
    description: string
    price: number
    stock: number
    reward_points: number
    roast_level: string
    taste: string
    is_active: boolean
    total_images: number
    created_at: string
    updated_at: string
    image_urls: string[]
}